Using click now...

+++ bs/utils.py
@@ -3,6 +3,8 @@ import os
 import re
 import shutil
 
+import bs
+
 def get_mtime(path):
     try:
         return os.path.getmtime(path)
@@ -11,10 +13,11 @@ def get_mtime(path):
         return -1
 
 def copy(source, destination):
-    if os.path.exists(source):
-        if get_mtime(destination) < get_mtime(source):
-            print('copying {} -> {}'.format(source, destination))
-            shutil.copy(source, destination)
+    if bs.CLEAN:
+        clean(destination)
+    elif os.path.exists(source) and get_mtime(destination) < get_mtime(source):
+        print('copying {} -> {}'.format(source, destination))
+        shutil.copy(source, destination)
 
 
 def touch(fname):
@@ -30,6 +33,7 @@ def touch(fname):
 def clean(fname):
     try:
         os.remove(fname)
+        print('removed {}'.format(fname))
     except:
         pass # who cares?
 
@@ -43,6 +47,7 @@ def clean(fname):
             break
         try:
             os.rmdir(dirname)
+            print('removed {}'.format(fname))
         except:
             break # there was a hidden file we didn't see
         dirname = os.path.dirname(dirname)
